0.6.1.0 Working towards submitting the integration into HA. This is the 2nd step.
Working towards submitting the integration into HA. This is the 2nd step. Auto code formatting has been done using black so it looks as if there are lots of code changes but in reality most are from the auto code formatter. This release does not call pyvisonic as a pypi library as I still need some flexibility to update the code. Some restructuring has been done, particularly with config variables, startup and restart.