0.8.1.0 - Updated Pin Code Settings and the Config for the Override Code
I have updated the logic around the following 3 configuration parameters: Override Code, Force Keypad and Arm Without User Code.
For most users you will not notice a difference, for the users that use these 3 settings in combination with the Mode of the Panel connection (Standard, Standard+ and Powerlink) you might notice a difference.  If you do notice a difference then I have updated note 2 on the wiki which describes the logic.

I have also changed the way the Override Code is set in the Configuration, it is now a number from 0 to 9999, 0 means do not use, numbers 1 to 9999 mean the 4 digit user code (miss out leading zeros i.e. code 123 means 0123).  This provides a small amount of validation checking.

I have also added a number validator for the Motion Off Delay (in seconds), it needs to be in the range 0 to 3000 seconds.