0.8.2.1 - Added 2 Powermaster sensors and changed checksum calc
There are 2 changes here:
- Added 2 Powermaster sensors to the list of recognised sensors
- Changed Checksum calc

For this latter change I noticed that while I've been using my PM30 panel that sometimes it would kick in the "bodge" that I use for checksum calculations on the messages.   This allows the checksum to be +/- 1 of the actual value. After investigating I have tweaked the checksum calc so it works properly and I also tried it on my main panel, a Powermax.  So, any problems then let me know in the usual way, either github or in the HA forum thread.