0.8.4.0 - Configuration and Event changes
Reintroduced the HA reload handler.

I have improved the removal and addition of the integration with HA, fixed config errors etc.

I have previously said that the integration will try to reconnect to the panel only once when the external physical connection is terminated (i.e. disconnected altogether such as a wifi router reboot). This happens when the disconnection occurs.
I have added some parameters for this to the configuration:
- Attempts to Try to Connect "Try this number of times to attempt the connection"
- Delay Between Attempts "Time delay (seconds) between reconnection attempts", this is use when the above setting is 2 or more.
These are described on the wiki in the table [here](https://github.com/davesmeghead/visonic/wiki#configuration-settings)
The defaults are set to do the same as the previous integration version i.e. "Attempts to Try to Connect" is set 1

Enabled "download diagnostics" on the settings/configuration.  Click the 3 vertical dots and you should see the menu item.  After you select it, HA gives you a text file that I can use to help with simple problems.
The diagnostics will give me the past 100 "things" that happened in the integration so I maye be able to help with simple problems.

For more difficult problem I'll still need a log file. If you have "logger" in your configuration.yaml you should also see a "Enable debug logging" on that same 3 dot menu.  If you have problems and can repeat it then enable debug logging, create the problem and then disable logging.  HA will then give you the log file as a text file.
These text files can help me to help you and they are really easy to create.

I have changed the wiki with lots of updates and restructure

For the following changes I suggest that you familiarise yourself with [this](https://github.com/davesmeghead/visonic/wiki#the-start-up-sequence-of-the-integration) on the wiki.

Note that in the following conditions the panel will go dormant:
- When it has tried the number of attempts to reconnect (as per the new parameters above) and failed
- When it has connected successfully but not received any data i.e. this is event condition=10 and please read the details for this
In these situations you will get a notification in the HA frontend only if you allow this in the configuration "List of Trigger for HA Notification" by including "Connection Problems".

If you use events from this integration then this is a breaking change!
- I have made all event control names lower case, with an added underscore to split words
- For condition=10 I have made these lowercase "NeverConnected", "DisConnected" to be consistent with other states.
- I have added "state" and "attempt" to condition 0.