0.9.3.0 - Panel 360, PanelConfig and EEPROM Decoding
Changes to accommodate a USB connection to a 360 Panel (and maybe a 360R Panel) It gets to Standard Plus OK but then fails to Enroll so I've added a PanelConfig setting to prevent it from Enrolling.
Extended the PanelConfig to ease the configuration checking. This includes making the keep alive interval set in PanelConfig per panel type.
Changed all EPROM to EEPROM for consistency. I'm not sure whether it is EPROM or EEPROM inside the panel but at least its consistent use now.
Further change to decoding the EEPROM
Various bug fixes and comments
Add the available property to Entities that didn't have it.