0.9.6.22 - Added X10 Home Assistant Service (Action) hassfest fix
0.9.6.22 - Added X10 Home Assistant Service (Action)
Although the X10 devices are Switches in Home Assistant, and provide a simple On/Off control in the Frontend, it is now possible to create a service (action) call to change the X10 device state as Off, On, Dimmer, Brighten. These last two have no feedback on how dim/bright the device is and so it cannot be made a Light Entity in Home Assistant.