0.12.0.0 - Significant update to the Integration to merge in Dev from past year or so
This is a significant update to the Integration and is likely to be a Breaking Change.
If you only use the Integrations Sensors and Alarm Entity in your Dashboards and Automations then you should be OK, but if you use any of their Attributes, or Actions / Events then you will need to update them.
If you have installation problems then I suggest that you delete the Integration from the Integrations Dashboard and Re-Create it.

There are significant improvement in Partitions, Language Support and many other things.

Improvements include...

The pyvisonic library does not manage the data to/from the panel, there is a separate comms manager that passes the data.  This means that the Entity State within HA can be kept whilst reconnecting the comms to the panel.

I have modified the reload and reconnect operations.  Definitions:
- Reconnect means that the HA Entity State has been kept and it is retrying to connect the comms to the panel. The Integration creates an HA Reconnect Action, this only operates once (per call) and simply tries to reconnect the comms i.e. the comms manager tries to reconnect.
- Reload is the HA Integration Reload operation i.e the whole Integration is reloaded, all HA State is reloaded.  You can use the built-in HA Action to do this.

If there is a disconnection then the integration uses the user config settings to determine what to do:
- reload count = 0 then do not automatically do anything.  You need to write your own automations. If there's a disconnection the integration will stop
- reload count > 0 then reload this many times with a time delay (user config) in between.  Note that reload count has a maximum of 1000000.

I have worked with Mark (link) to support the use of this Integration with a Powerlink hardware and a remote connection.

There is a new HA Action to retrieve sensor and x10 information, the action does not actually do anything else other then retrieve data. However, note that the Command Action (e.g. Arm, Disarm) also returns this information.

For some PowerMaster panel wireless sensors, the Integration may be able to obtain additional attributes to provide:
- Missing, wireless one-way or inactive indications. This gives a status of the wireless connectivity of the sensor.
- Temperature and Luminance for supported sensors. This is experimental as I have no way to test these.

Sensors have 2 Zone Name attributes:
- The name translated by the HA language translation files and
- The zone name that is reported by the Panel itself.

I have removed the 'lastevent' attribute from Alarm Entity.  I did say I'd do this eventually!  Please use lasteventname and lasteventaction instead.

I have renamed the 'mode' attribute of the Alarm Entity.  This seemed to be in conflict with HA somehow so it is now called 'emulationmode'.

PowerMaster Panels attempt to retrieve the panel data using B0 messages first, if the does not succeed then it tries to download the EEPROM. It is like this as the B0 message download takes a fraction of the time compared to EEPROM.

If the data from the Panel says that it Supports a PGM (regardless of whether the actual hardware exists) then a Home Assistant Switch is created. Button 3 and 6 on the panel usually control the PGM on/off. You can use these from your panel to set the HA Switch state.

Lots of Fixes to many things, lots of code tidying.